Remove i18n functionality
I18n functionality is a little overkill for this little project right
now, so this commit zaps all related files from the template.

Additionally, we also replace the lodash.merge dependency with a
dependency on lodash as this was breaking the dist/production build with
"Error: Cannot find module 'lodash'". I have no idea why, but replacing
lodash.merge (which was pulled in by the I18n deps) with lodash seems to
have fixed it 🤷