Ignore electron imports
Currently, eslint complains at every import from electron as follows:
```
'electron' should be listed in the project's dependencies, not devDependencies
```

I think the reasoning here is that if we're importing from some pacakge in our production code, it shouldn't be listed as a devDependency. That assumption isn't really valid for Electron though. Fixed by telling eslint to conisder electron as a core module as described in [this github issue](https://github.com/SimulatedGREG/electron-vue/issues/423)