Add MaterialUI and get it working
This commit adds in the MaterialUI dependency and includes a *lot* of
fiddling to get MaterialUI components to render properly. The hackiest
thing here is hardcoding a nonce to appease our Content Security Policy
settings. 😕

In any case, it works for now. Should probably find a better way to do
that though: insert a placeholder nonce (e.g. {{nonce_value}}) and then
replace it by running the generated html through a templating engine
like ejs or sth.