Add VideoUtils module + preload and main bindings
The new VideoUtils module wraps youtubedl and exposes 2 methods
mainBindings() and preloadBindings() which are then used in the main.js
(main process) and preload.js (renderer preload script). These *binding
methods set up the necessary ipcMain and ipcRenderer inter-process
communication(ipc) signals to allow us to easily download videos 'from'
the renderer process. We need to do it this way (expose the
functionality through the preload script) because the renderer processes
run in a sandbox and cannot use the required node functionality.