Return thumbnails as base64str images
Electron won't let us load local images in a renderer process without a whole lot of hoopla. To get around that, this commit just encodes the image as a single nase64 string that we can pass directly from the main process to the renderer process.

To help with this process, we also add in a new dependency on
image-to-base64 which does the conversion for us.