Download youtube-dl on first load
This pushes the responsibility of deciding how/which youtube-dl binary
to download into the node-youtube-dl library which already has logic
around that. After this commit, on first load, the app will download the
required youtube-dl binary for the machine on which it's running. On
subsequest uses, if the binary is already present, it just skips that
step.